refactor(alertd): move the doctor subsystem into the alertd crate
Relocate the doctor framework, checks, sweep orchestration, and the
DoctorTask from bestool-tamanu and bestool into bestool-alertd. alertd
now owns the monitoring engine and depends on bestool-tamanu for common
Tamanu domain code; bestool's doctor CLI keeps only arg parsing,
rendering, and daemon-fetch, calling into bestool_alertd::doctor.

This is a behaviour-preserving relocation. The one functional fix:
perform_sweep and DoctorTask take the running binary's version as a
parameter, so the wire payload reports bestool's version rather than
alertd's (env!("CARGO_PKG_VERSION") would otherwise resolve to the
library crate).
