+++ crates/alertd/Cargo.toml
@@ -19,15 +19,20 @@ required-features = ["cli"]
 [dependencies]
 axum = "0.8.9"
 bestool-canopy = { version = "0.2.0", path = "../canopy" }
+bestool-kopia = { version = "0.1.0", path = "../kopia" }
 bestool-postgres = { version = "1.0.11", path = "../postgres" }
+bestool-tamanu = { version = "0.10.2", path = "../tamanu" }
 blake3 = "1.8.5"
 bytes = "1.9.0"
 clap = { workspace = true, optional = true, features = ["env", "wrap_help"] }
 clap-markdown = { version = "0.1.5", optional = true }
 dirs = "6.0.0"
+duct = "1.1.0"
 futures = { workspace = true }
 glob = "0.3.3"
+hickory-resolver = "0.26.1"
 jiff = { version = "0.2.24", features = ["serde"] }
+node-semver = "2.2.0"
 lloggs = { workspace = true, optional = true }
 mailgun-rs = "2.0.2"
 miette = { workspace = true }
@@ -46,6 +51,7 @@ tera = "1.20.0"
 thiserror = { workspace = true }
 tokio = { workspace = true, features = ["full"] }
 tokio-postgres = { version = "0.7.17", features = ["with-jiff-0_2", "with-serde_json-1"] }
+tokio-stream = "0.1.17"
 tokio-util = { workspace = true }
 tower = "0.5.2"
 tower-http = { version = "0.6.6", features = ["trace"] }
