+++ crates/alertd/src/doctor.rs
@@ -1,0 +1,9 @@
+pub mod check;
+pub mod checks;
+pub mod progress;
+pub mod server_info;
+pub mod sweep;
+pub mod task;
+
+pub use sweep::{SweepResult, overall_from_payload, perform_sweep};
+pub use task::DoctorTask;
