+++ crates/alertd/src/doctor/check.rs
