+++ crates/alertd/src/doctor/checks/db_connect.rs
