+++ crates/alertd/src/doctor/checks/http_errors.rs
