+++ crates/alertd/src/doctor/checks/memory.rs
