+++ crates/alertd/src/doctor/checks/migrations.rs
