+++ crates/alertd/src/doctor/checks/sync_sessions.rs
