+++ crates/alertd/src/doctor/checks/uptime.rs
