+++ crates/alertd/src/doctor/server_info.rs
@@ -12,7 +12,7 @@ use sysinfo::{Disks, System};
 use tokio::net::TcpStream;
 use tracing::debug;
 
-use crate::server_info::{detect_node_version, detect_virtualisation};
+use bestool_tamanu::server_info::{detect_node_version, detect_virtualisation};
 
 const PROBE_TIMEOUT: Duration = Duration::from_secs(3);
 const IPV4_PROBE_ADDR: SocketAddr = SocketAddr::new(IpAddr::V4(Ipv4Addr::new(1, 1, 1, 1)), 443);
@@ -38,7 +38,7 @@ pub struct Filesystem {
 #[derive(Debug, Clone, Serialize)]
 #[serde(rename_all = "camelCase")]
 pub struct ServerInfo {
-	pub bestool_version: &'static str,
+	pub bestool_version: String,
 	pub tamanu_version: String,
 	/// Host's installed Node.js version (bare, no leading `v`), if node is on
 	/// `PATH`. Omitted when node isn't installed or can't be queried.
@@ -92,17 +92,11 @@ pub struct ServerFacts {
 
 /// Build the status-payload `ServerInfo` block.
 ///
-/// `bestool_version` is the version of the *calling* bestool binary — it must
-/// be provided by the caller (typically `env!("CARGO_PKG_VERSION")` resolved
-/// in the bestool crate) rather than evaluated here, because this code lives
-/// in the `bestool-tamanu` library crate, where `env!("CARGO_PKG_VERSION")`
-/// would resolve to that crate's own version (`bestool-tamanu`, currently 0.1.x)
-/// — the bug this signature change fixes.
-pub async fn gather(
-	bestool_version: &'static str,
-	tamanu_version: &str,
-	facts: ServerFacts,
-) -> ServerInfo {
+/// `bestool_version` is the version of the *calling* binary — it must be
+/// provided by the caller (`env!("CARGO_PKG_VERSION")` resolved in the bestool
+/// crate) rather than evaluated here, since in this library crate it would
+/// resolve to the library's own version instead of the running binary's.
+pub async fn gather(bestool_version: &str, tamanu_version: &str, facts: ServerFacts) -> ServerInfo {
 	let disks = Disks::new_with_refreshed_list();
 	let filesystems = disks
 		.iter()
@@ -122,7 +116,7 @@ pub async fn gather(
 		.map(|s| s.to_string());
 
 	ServerInfo {
-		bestool_version,
+		bestool_version: bestool_version.to_string(),
 		tamanu_version: tamanu_version.to_string(),
 		node_version: detect_node_version().await,
 		hostname: System::host_name(),
