--- crates/tamanu/src/doctor.rs
@@ -1,4 +1,0 @@
-pub mod check;
-pub mod checks;
-pub mod progress;
-pub mod server_info;
