+++ crates/tamanu/src/lib.rs
@@ -18,9 +18,6 @@ pub mod versions;
 
 pub mod systemd;
 
-#[cfg(feature = "doctor")]
-pub mod doctor;
-
 /// What kind of server to interact with.
 #[derive(Copy, Clone, Debug, PartialEq, Eq, PartialOrd, Ord)]
 pub enum ApiServerKind {
